- Отрефакторила модели базы данных. Убрала camelCase в иненовании столбцов; - Дорабатывала фильты; - Добавила проверки  на редактирование и удаление тэгов, если они есть на других страницах. Теперь всплывает диалоговое окно для подверждения; - Добавила todo  для фикса багов и вычистки кода, которые буду делать в следующих коммитах;

+++ backend/src/components/Tag/TagController.ts
@@ -2,7 +2,8 @@ import {Request, Response} from 'express';
 import {Tag} from '../../models/Tag.model';
 import {Page} from '../../models/Page.model';
 import {PageTag} from "../../models/PageTag.model";
-import {Album} from "../../models/Album.model";
+
+// todo Cелать "не прямое удаление" тэга со страницы и в модель внести соответствующие поля
 
 export class TagController {
     public addTag(req: Request, res: Response) {
@@ -18,25 +19,47 @@ export class TagController {
         }).catch((err: Error) => res.status(500).json(err));
     }
 
+    private async _checkTag(tagId: number) {
 
-    private _checkTag() {
-        console.log(' Нужно проверить тэг');
+        if (tagId) {
+
+            return await PageTag.count({
+                where: {tag_id: [tagId]}
+            }).then(dataCount => {
+
+                console.log('dataCount');
+                console.log(dataCount);
+                return (dataCount <= 1);
+            }).catch
+            (
+                () => {
+                    return false;
+                }
+            );
+        } else {
+            return false;
+        }
     }
 
-    public editTag(req: Request, res: Response) {
+    public async editTag(req: Request, res: Response) {
 
-        // todo Проверять, присутствует ли тэг  на других страницах. Если да, то редактировать нельзя
-        // todo Возможно не плохо бы знать айдишник страницы на которой происходит действо
+        let isDataCountValid = await this._checkTag(req.body.tagId);
 
-        this._checkTag();
+        if (isDataCountValid || req.body.isEditingApproved) {
 
-        Tag.update(
-            {
-                title: req.body.title,
-            },
-            {where: {id: req.body.tagId, type: req.body.tagType}}
-        ).then(() => res.status(202).json({data: "success"}))
-            .catch((err: Error) => res.status(500).json(err));
+            Tag.update(
+                {
+                    title: req.body.title,
+                },
+                // todo  А нужно ли реально tagType? Следующим коммитом хорошенько проверить
+                {where: {id: req.body.tagId, type: req.body.tagType}}
+            ).then(
+                () => res.status(202).json({data: "success"})
+            ).catch((err: Error) => res.status(500).json(err));
+
+        } else {
+            res.status(403).json({data: "Нельзя редактировать тэг"});
+        }
     }
 
     public getTags(req: Request, res: Response) {
@@ -60,32 +83,51 @@ export class TagController {
         }
     }
 
-    public deleteTag(req: Request, res: Response) {
+    public getTagsByPageType(req: Request, res: Response) {
 
-        const id = req.params.id;
+        Tag.findAll<Tag>({
+            where: {
+                type: req.params.tagType,
+                page_type: req.params.pageType,
+            }
+        }).then(tag => {
+            console.log(tag);
+            res.json(tag);
+        }).catch((err: Error) => res.status(500).json(err));
 
-        // todo Проверять, присутствует ли тэг  на других страницах. Если да, то удалять нельзя
-        // todo Возможно не плохо бы знать айдишник страницы на которой происходит действо
+    }
 
-        Tag.destroy({
-            where: {id: id}
-        })
-            .then(num => {
-                if (num == 1) {
-                    res.send({
-                        message: "Tag was deleted successfully!"
-                    });
-                } else {
-                    res.send({
-                        message: `Cannot delete tag with id=${id}. Maybe tag was not found!`
-                    });
-                }
+    public async deleteTag(req: Request, res: Response) {
+
+        const isDeletingApproved = JSON.parse(req.params.isDeletingApproved);
+        const tagId = Number(req.params.id);
+
+        let isDataCountValid = await this._checkTag(tagId);
+
+        if (isDataCountValid || isDeletingApproved) {
+
+            Tag.destroy({
+                where: {id: tagId}
             })
-            .catch(err => {
-                res.status(500).send({
-                    message: "Could not delete tag with id=" + id
+                .then(num => {
+                    if (num == 1) {
+                        res.send({
+                            message: "Tag was deleted successfully!"
+                        });
+                    } else {
+                        res.send({
+                            message: `Cannot delete tag with id=${tagId}. Maybe tag was not found!`
+                        });
+                    }
+                })
+                .catch(err => {
+                    res.status(500).send({
+                        message: "Could not delete tag with id=" + tagId
+                    });
                 });
-            });
+        } else {
+            res.status(403).json({data: "Cannot delete tag"});
+        }
     }
 
 }
